Renaming files to comply with the style guide.
