+++ src/meson.build
@@ -8,7 +8,7 @@ wayfire_sources = ['main.cpp',
                    'core/core.cpp',
                    'core/img.cpp',
                    'core/wm.cpp',
-                   'core/view_access_interface.cpp',
+                   'core/view-access-interface.cpp',
 
                    'core/seat/pointing-device.cpp',
                    'core/seat/input-manager.cpp',
@@ -91,7 +91,7 @@ install_headers(['api/wayfire/compositor-surface.hpp',
                  'api/wayfire/surface.hpp',
                  'api/wayfire/view-transform.hpp',
                  'api/wayfire/view.hpp',
-                 'api/wayfire/view_access_interface.hpp',
+                 'api/wayfire/view-access-interface.hpp',
                  'api/wayfire/workspace-manager.hpp',
                  'api/wayfire/workspace-stream.hpp',
                  'api/wayfire/input-device.hpp',
